+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### v6.0.0
 
-- Supported Node.js versions: `^22.19.0 || ^24.0.0`.
+- Supported Node.js versions: `^22.19.0 || ^24.0.0`;
+- Supported Zod version: `^4.3.4`;
 - The `typescript` dependency is now optional and only required for making `Integration`:
   - Either import and assign the `typescript` property to its constructor argument;
   - Or use the new static async method `create()` to delegate the import;